Add Juliet autoevaluator
This commit also bumps the size of interned strings up, since the juliet
evaluation uses more than 256 strings.

+++ src/interned_string.rs
@@ -8,12 +8,12 @@ lazy_static! {
 
 #[derive(Debug, Clone, Hash, Ord, PartialOrd, Eq, PartialEq, Copy)]
 pub struct InternedString {
-    index: u8,
+    index: u16,
 }
 
 impl InternedString {
     fn max() -> usize {
-        ::std::u8::MAX as usize
+        ::std::u16::MAX as usize
     }
     pub fn from_string(s: &str) -> Self {
         let mut intern = STRING_INTERN.lock().unwrap();
@@ -23,7 +23,7 @@ impl InternedString {
         assert!(intern.0.len() < InternedString::max());
         intern.0.push(s.to_string());
         let out = Self {
-            index: (intern.0.len() - 1) as u8,
+            index: (intern.0.len() - 1) as u16,
         };
         intern.1.insert(s.to_string(), out);
         out
